Organize scene files
